mv some func to np_ script
